backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
redis==5.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
openai>=2.0.0
pydantic==2.9.2
python-multipart==0.0.12
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
